% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_deepviz.R
\name{plot_deepviz}
\alias{plot_deepviz}
\title{Plot deepviz}
\usage{
plot_deepviz(n)
}
\arguments{
\item{n}{A numeric vector with an element for each layer that indicates the number of nodes in that layer.}
}
\description{
Plot deepviz
}
\examples{
nn <- c(4, 10, 3)

plot_deepviz(nn)
}
