% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_model.R
\name{plot_model}
\alias{plot_model}
\title{Plot keras model}
\usage{
plot_model(model)
}
\arguments{
\item{model}{A keras model defined using \link[keras:keras_model_sequential]{keras::keras_model_sequential}}
}
\description{
Plot keras model
}
\examples{
require(keras)

model <- keras_model_sequential() \%>\%
  layer_dense(10, input_shape = c(64, 64)) \%>\%
  layer_conv_1d(filters = 16, kernel_size = 8) \%>\%
  layer_max_pooling_1d() \%>\%
  layer_flatten() \%>\%
  layer_dense(25) \%>\%
  layer_dense(25, activation = "relu") \%>\%
  layer_dropout(0.25) \%>\%
  layer_dense(2, activation = "sigmoid")


model

model \%>\% plot_model()

}
