% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_model.R
\name{plot_model}
\alias{plot_model}
\title{Plot keras model}
\usage{
plot_model(model)
}
\arguments{
\item{model}{A \link[keras:keras_model_sequential]{keras::keras_model_sequential} model}
}
\description{
Plot keras model
}
\examples{
require(keras)

model <- keras_model_sequential() \%>\%
  layer_dense(10, input_shape = 4) \%>\%
  layer_dense(25, input_shape = 4) \%>\%
  layer_dense(25, input_shape = 4, activation = "relu") \%>\%
  layer_dropout(0.25) \%>\%
  layer_dense(2, activation = "sigmoid")

class(model)

model \%>\% plot_model()


}
